pyproject.toml
[project]
name = "google-ads-mcp"
version = "0.1.0"
description = "MCP server for Google Ads API integration"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmcp>=2.12.4",
    "pydantic-settings>=2.11.0",
    "google-ads>=28.0.0",
    "google-auth>=2.0.0",
    "google-auth-oauthlib>=1.0.0",
]


[project.scripts]
google-ads-mcp = "google_ads_mcp.__main__:main"
